+++ crates/apimock-routing/src/rule_set/rule/when/request/body.rs
@@ -88,3 +88,6 @@ impl std::fmt::Display for Body {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;