+++ crates/apimock-routing/src/rule_set/rule/when/request/headers.rs
@@ -73,3 +73,6 @@ impl std::fmt::Display for Headers {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;